refactor: 추천 이벤트 배너 슬라이드 컴포넌트들 사용하여 구현
