+++ server/data/webtoon.json
@@ -109,7 +109,7 @@
                 "updateDay": "화"    
             }
         ],
-        "recommendEventTop": [
+        "recommendEvent": [
             {
                 "index": 0,
                 "src": "https://dn-img-page.kakao.com/download/resource?kid=DHuH7/hzp2gCacSm/CxKcdmOaKcmNKqULsxPtm0"
